% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPlot.R
\name{gPlot}
\alias{gPlot}
\title{gPlot}
\usage{
gPlot(fit_obj, spike_param)
}
\arguments{
\item{fit_obj}{Fitted object from function 'MUSS'.}

\item{spike_param}{The value of the spike parameter that needed to be specified,
which must be in \code{spike_params} designated in \code{MUSS} function.
Each \code{spike_param} corresponds to a unique gPlot.
If default \code{spike_params} is used in \code{MUSS}, one can check the values of
\code{spike_params} by calling \verb{$spike_params}
from the returned object.}
}
\description{
Plot \eqn{g(\bold{\beta^{(t+1)}},\theta^{(t+1)},\sigma^{(t+1)}|\bold{\beta^{(t)}},\theta^{(t)},
\sigma^{(t)})} over iterations at a specified spike parameter.
\eqn{g(\bold{\beta},\theta,\sigma|\bold{\beta^{(t)}},\theta^{(t)},\sigma^{(t)})} is
the expected value of the log likelihood function with respect to the latent
variables \eqn{X} and \eqn{\bold{\gamma}}, conditioning on current
estimated parameters.
}
